feat: implement order management system with CRUD operations and stat…

+++ src/main/java/spring/practice/elmenus_lite/repository/CustomerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import spring.practice.elmenus_lite.model.CustomerModel;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<CustomerModel, Long> {
+public interface CustomerRepository extends JpaRepository<CustomerModel, Integer> {
 }
